More fixed on pages show and pages index

+++ frontend/components/pages/page_index_item.jsx
@@ -20,29 +20,30 @@ class PageIndexItem extends React.Component {
   }
 
   componentDidMount() {
-    // this.props.fetchPage(this.props.page.id);
-    this.props.fetchPlaces();
     this.props.fetchPages(this.props.match.params.placeId);
   }
 
+  componentWillUnmount() {
+    // this.props.fetchPages([]);
+  }
+
   sendData(e) {
+    // console.log('ITEMS DATA')
     // console.log(this.props)
-    // this.props.history.push(`/places/${this.props.match.params.placeId}/pages/${this.props.pageId}/`);
-    // <NavLink to={`/places/${this.props.match.params.placeId}/pages/${this.props.pageId}/`} />;
-    console.log('ITEMS DATA')
-    console.log(this.props)
-    // const { childrenTest } = this.props;
-    // console.log(childrenTest)
-    // console.log(this.props.page)
-    // console.log(e.currentTarget)
   }
 
   destroyPage(e) {
     e.preventDefault();
     e.stopPropagation();
-    if (this.props.page.id === this.props.pageId) {
-      this.props.deletePage(this.props.page.id);
-      this.props.history.push(`/places/${this.props.placeId}/pages/${this.props.thisPlace.pages[0].id}`);
+    if (this.props.currentUser.id === this.props.thisPlace.owner_id) {
+        // || this.props.currentUser.id === this.props.pages[this.props.match.params.pageId].owner_id) {
+      if (this.props.page.id === this.props.pageId) {
+        this.props.deletePage(this.props.page.id)
+          .then(() => this.props.history.push(`/places/${this.props.placeId}/pages/${this.props.thisPlace.pages[0].id}`));
+        // let firstPage = Object.values(this.props.thisPlace).pages[0];
+        // this.props.history.push(`/places/${this.props.placeId}/pages/${firstPage}`);
+  
+      }
 
     }
   }
@@ -64,45 +65,27 @@ class PageIndexItem extends React.Component {
   createNewPage(e) {
     e.preventDefault();
     const { placeId, page } = this.props.match.params;
-    // const nextPage = Object.values(getState().entitites.pages)[Object.values(getState().entities.pages).length].id + 1;
     let defaultPage = {
       title: 'New Page',
       place_id: placeId,
       parent_page_id: this.props.page.id,
     };
-    this.props.createPage(defaultPage);
-    // this.forceUpdate();
-    // .then(() => this.props.history.push(`/places/${placeId}/pages/${Object.values(getState().entitites.pages)[Object.values(getState().entities.pages).length].id}`));
-    // this.props.history.push(`/places/${placeId}/pages/${newPageId}`);
-    
-    // this.props.createPage(defaultPage);
-    // const newPageId = Object.values(getState().entitites.pages)[Object.values(getState().entities.pages).length].id;
-    // this.props.history.push(`/places/${placeId}/pages/${newPageId}`);
-
-    // console.log('CREATING NEW PAGE')
-    // console.log(this.props)
-    // console.log('WOO YEAH')
-    // console.log(this.props.pages);
+    this.props.createPage(defaultPage)
+      .then((response) => console.log(response));
   }
 
   componentDidUpdate(prevProps) {
     if (Object.values(getState().entities.pages).length > Object.values(prevProps.pages).length) {
-      console.log('YEAH');
-      // console.log(this.props.pages);
+      // console.log('YEAH');
       const length = Object.values(this.props.pages).length;
       const newPageId = Object.values(this.props.pages)[length - 1].id;
-      // console.log(this.props)
-      // console.log(getState().entities.pages)
-      // newPageId = Object.values(getState().entitites.pages)[Object.values(getState().entities.pages).length].id;
-      this.props.history.push(`/places/${this.props.placeId}/pages/${newPageId}`);
-      this.forceUpdate();
+      // this.props.history.push(`/places/${this.props.placeId}/pages/${newPageId}`);
       // this.forceUpdate();
     }
   }
 
   render () {
     const {title, id, src, layers, classTitle, placeId, page } = this.props; 
-    // const children = this.props.page.children;
     const children = this.props.page === undefined ? '' : this.props.page.children;
 
     let styles = {
@@ -112,7 +95,6 @@ class PageIndexItem extends React.Component {
     return (
       <ul>
       <li className={classTitle} onClick={this.sendData}>
-        {page ? console.log('page test') : ''}
         <Link to={`/places/${this.props.match.params.placeId}/pages/${this.props.pageId}/`}>
           <img className="page-index-nav-icon" src={src}></img>
           <span className="pages-index-selected">{title}</span>
